Ensure m/bcast settings are on when macvlan and mac callbacks are received.
In the prototype the parameters for every VF were pushed to the NIC
with each callback.  When moving VFd toward a prodcution piece of
code the push of NIC parms for a VF were limited only to when the
device was reset, the VF went down and back up, or was added to the
view of VFd.  Inadvertantly, the u/m/bcast flags were not being reset
and it seems they need to be when we receive macvlan and vlan callback
messages.