Tweak push of VF parms to NIC on callback.
The parms must be pushed to the NIC immediately on a negotiate api
callback or issues arise when a machine is directly connected (no
switch) and that is being used for sr-iov on both sides.  All other
callbacks are now queuing a reset to happen when the queue for the
VF goes hot.  That will be nearly immediately if the queue was not
stopped when the callback was received, but should allow for timing
issues to be avoided if it is.