Adding mac list support for mlx5 devices
This change adds mac whitelist support in mlx5 device by
allowing user to provide a list of mac in the vf cfg file.

The first mac in the list will be set as the default mac of the vf
and the rest will be added to the whitelist.

Issue: None.
Change-Id: I4447b876b0ab71fd5ee1d57fe2391f01bd3dd9c5
Signed-off-by: Ariel Levkovich <lariel@mellanox.com>